requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.1
django-cors-headers==2.4.0
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
six==1.11.0
